feat/JUS-90: Finished creating new UI for User login
- Created new UI for user login page

- Deleted old page and related files, will need to setup the form and connect to backend
