+++ client/package.json
@@ -27,7 +27,7 @@
     "eject": "react-scripts eject",
     "serverstart": "nodemon ../server/app.js"
   },
-  "proxy": "http://localhost:3000",
+  "proxy": "http://127.0.0.1:3005/",
   "eslintConfig": {
     "extends": [
       "react-app",
