+++ client/public/stylesheet.css
@@ -1,0 +1,4 @@
+html *
+{
+    font-family: "Droid Sans Mono Dotted" !important;
+}
