+++ client/src/index.js
@@ -2,13 +2,13 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import reportWebVitals from './reportWebVitals'
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
-import Login from './pages/login/login.js'
-import Signup from './pages/customer/signup/signup.jsx'
+import Login from './pages/user/login/login.jsx'
+import Signup from './pages/user/signup/signup.jsx'
 import OrganizerRegister from './pages/organizer/register.jsx'
 import OrgVerifyEmail from './pages/organizer/verifyEmail/verifyEmail.jsx'
 import OrgVerificationRequired from './pages/organizer/verificationRequired/verificationRequired.jsx'
-import CustomerVerifyEmail from './pages/customer/verifyEmail/verifyEmail.jsx'
-import CustomerVerificationRequired from './pages/customer/verificationRequired/verificationRequired.jsx'
+import CustomerVerifyEmail from './pages/user/verifyEmail/verifyEmail.jsx'
+import CustomerVerificationRequired from './pages/user/verificationRequired/verificationRequired.jsx'
 import Header from './components/header/Header'
 import { store } from './store.js'
 import {Provider} from 'react-redux'
