+++ client/src/pages/user/verificationRequired/verificationRequired.jsx
