+++ client/src/pages/user/verifyEmail/verifyEmail.jsx
